Working on new file structure to test it out.
